# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sales API",
        "version": "1.0.0"
    },
    "paths": {
        "/v1/products/{product_name}/dates": {
            "get": {
                "summary": "Get all dates a product was sold",
                "parameters": [
                    {
                        "name": "product_name",
                        "in": "path",
                        "description": "The name of the product",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "format": "date"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/v1/google/auth": {
            "get": {
                "summary": "Google OAuth2 authorization",
                "description": "Redirects the user to the Google OAuth2 authorization page.",
                "responses": {
                    "default": {
                        "description": "Successful response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/google/callback": {
            "get": {
                "summary": "Google OAuth2 callback",
                "description": "Handles the Google OAuth2 callback request and saves the user data to the database.",
                "responses": {
                    "default": {
                        "description": "Successful response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/register": {
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "The user to register",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "summary": "Log in a user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "The user to log in",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged in successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/v1/sales/{year}/{month}": {
            "get": {
                "summary": "Get sales report for a given year and month",
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "description": "The year of the sales data",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "month",
                        "in": "path",
                        "description": "The month of the sales data",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SalesReport"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/v1/sales": {
            "post": {
                "summary": "Add a new sales record",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "The sales record to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SalesRecord"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Sales record added successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/v1/products/{product_name}": {
            "delete": {
                "summary": "Delete all mentions of a product",
                "parameters": [
                    {
                        "name": "product_name",
                        "in": "path",
                        "description": "The name of the product",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "SalesReport": {
            "type": "object",
            "properties": {
                "product": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "format": "int32"
                },
                "total": {
                    "type": "number",
                    "format": "float"
                }
            }
        },
        "SalesRecord": {
            "type": "object",
            "properties": {
                "product": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "format": "int32"
                },
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "total": {
                    "type": "number",
                    "format": "float"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}